refactor(web): consolidate BlogPost and ProjectPost types into a single post.ts file (#1370)
